+++ storage/src/object/mod.rs
@@ -1,4 +1,15 @@
-pub mod object;
+extern crate serde;
+extern crate serde_json;
 
 #[cfg(test)]
 mod tests;
+
+// Object stores the basic object that all game objects willb e serialized 
+// into.
+#[derive(Serialize, Deserialize, Debug, PartialEq)]
+pub struct Object {
+    pub id: u64,
+    pub name: String,
+    pub description: String,
+    pub contents: Vec<u64>,
+}
